Add files via upload
Barrel is a 14mm PETG PC watercooling pipe. Machine the ends down to an OD of 0.495 in and cut the threads single point or with a 1/2-27 die. Rod is a 2mm titanium rod. Use 004 o-rings and 007 for the piston. Print with supports and pick them out with a blade. Locktite/superglue the rod tip on, test thread into the piston and section, and then cut the rod to length and assemble, gluing the knob onto the rod.